+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueScreen.kt
@@ -1,7 +1,6 @@
 package org.simple.clinic.home.overdue
 
 import android.content.Context
-import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -150,11 +149,7 @@ class OverdueScreen : BaseScreen<
 
   override fun createUpdate(): Update<OverdueModel, OverdueEvent, OverdueEffect> {
     val date = LocalDate.now(userClock)
-    val canGeneratePdf = Build.VERSION.SDK_INT >= Build.VERSION_CODES.N
-    return OverdueUpdate(
-        date = date,
-        canGeneratePdf = canGeneratePdf
-    )
+    return OverdueUpdate(date = date, canGeneratePdf = true)
   }
 
   override fun createInit() = OverdueInit()