+++ app/src/main/java/org/simple/clinic/medicalhistory/MedicalHistory.kt
@@ -22,6 +22,7 @@ import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HasHadAStroke
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsOnDiabetesTreatment
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsOnHypertensionTreatment
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsSmoking
+import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsUsingSmokelessTobacco
 import org.simple.clinic.patient.PatientUuid
 import org.simple.clinic.patient.SyncStatus
 import java.time.Instant
@@ -57,6 +58,8 @@ data class MedicalHistory(
 
     val isSmoking: Answer,
 
+    val isUsingSmokelessTobacco: Answer,
+
     @ColumnInfo(name = "cholesterol_value")
     val cholesterol: Float?,
 
@@ -89,6 +92,7 @@ data class MedicalHistory(
       is IsOnHypertensionTreatment -> copy(isOnHypertensionTreatment = answer)
       IsOnDiabetesTreatment -> copy(isOnDiabetesTreatment = answer)
       IsSmoking -> copy(isSmoking = answer)
+      IsUsingSmokelessTobacco -> copy(isUsingSmokelessTobacco = answer)
     }
   }